Move Python Runtime Extensions
Move the Python runtime extensions to a subfolder `extensions/` to
avoid future naming collisions with types we actually define from
pybind11.
